dvi/dvi_modeline/tmds: restructure file/naming heirarchy
