+++ gateware/src/tiliqua/dvi.py
@@ -1,140 +1,120 @@
-# DVI PHY implementation (TMDS Encoder and DDR driver)
-#
-# This is an Amaranth rewrite of the DVI core in:
-# "Project F Library - TMDS Encoder for DVI"
-#   - Original attribution:
-#       Copyright Will Green
-#       Open source hardware released under the MIT License
-#       Learn more at https://projectf.io
-
 from amaranth import *
 from amaranth.lib import wiring
 from amaranth.lib.wiring import In, Out
 
+from vendor  import tmds
 from tiliqua import sim
 
-class TMDSEncoder(wiring.Component):
-    """
-    TMDS (Transition Minimized Differential Signaling) Encoder for DVI.
-    This component encodes 8-bit color data and 2-bit control data into
-    10-bit TMDS symbols according to the DVI specification.
+class DVITimingGen(wiring.Component):
 
-    This component operates in a `dvi` domain running at the pixel clock.
+    """
+    State machine to generate pixel position and hsync/vsync/de signals,
+    based on a DVI timing modeline that may be changed at runtime.
     """
 
-    data_in: In(8)    # Color data
-    ctrl_in: In(2)    # Control data
-    de: In(1)         # Data enable
-    tmds: Out(10)     # Encoded TMDS data
+    class TimingProperties(wiring.Signature):
+        """
+        Signals used to relay properties of current DVI modeline.
+        May be changed at runtime, as long as the DVI cores are in reset.
+        """
+        def __init__(self, coord_width=12):
+            super().__init__({
+                "h_active":      Out(coord_width),
+                "h_sync_start":  Out(coord_width),
+                "h_sync_end":    Out(coord_width),
+                "h_total":       Out(coord_width),
+                "h_sync_invert": Out(1),
+                "v_active":      Out(coord_width),
+                "v_sync_start":  Out(coord_width),
+                "v_sync_end":    Out(coord_width),
+                "v_total":       Out(coord_width),
+                "v_sync_invert": Out(1),
+                "active_pixels": Out(coord_width*2),
+            })
 
-    def elaborate(self, platform):
+    class ControlSignals(wiring.Signature):
+        """
+        Outgoing display synchronization signals generated by this core.
+        """
+        def __init__(self):
+            super().__init__({
+                "hsync": Out(1),
+                "vsync": Out(1),
+                "de":    Out(1),
+            })
+
+    def __init__(self):
+        super().__init__({
+            "timings": In(self.TimingProperties()),
+            # Control signals without inversion applied.
+            # Useful for driving logic external to this core (e.g. do something
+            # based on vsync, without having to worry the display might invert it).
+            "ctrl": Out(self.ControlSignals()),
+            # Control signals with inversion applied.
+            "ctrl_phy": Out(self.ControlSignals()),
+        })
+
+    def elaborate(self, platform) -> Module:
         m = Module()
 
-        # Register for output
-        tmds_r = Signal(10, init=0b1101010100) # equivalent to ctrl 2'b00
-        m.d.comb += self.tmds.eq(tmds_r)
+        timings = self.timings
 
-        # Register for ongoing DC bias
-        bias = Signal(signed(5), init=0)
+        # coordinates at which each frame starts
+        self.h_reset = timings.h_active - timings.h_total
+        self.v_reset = timings.v_active - timings.v_total
 
-        # Select basic encoding based on number of ones in the input data
-        data_1s = Signal(4)
-        use_xnor = Signal()
+        # coordinates at which sync starts / ends
+        self.hs_start = timings.h_sync_start - timings.h_total
+        self.vs_start = timings.v_sync_start - timings.v_total
+        self.hs_end   = timings.h_sync_end   - timings.h_total
+        self.vs_end   = timings.v_sync_end   - timings.v_total
 
-        # Calculate number of ones in data_in (manually unrolled for clarity)
-        m.d.comb += data_1s.eq(
-            self.data_in[0] + self.data_in[1] + self.data_in[2] + self.data_in[3] +
-            self.data_in[4] + self.data_in[5] + self.data_in[6] + self.data_in[7]
-        )
+        # current position
+        self.x = Signal(signed(12))
+        self.y = Signal(signed(12))
 
-        # Determine encoding type
-        m.d.comb += use_xnor.eq((data_1s > 4) | ((data_1s == 4) & (self.data_in[0] == 0)))
+        with m.If(ResetSignal("sync")):
+            m.d.dvi += [
+                self.x.eq(self.h_reset),
+                self.y.eq(self.v_reset),
+            ]
+        with m.Elif(self.x == (timings.h_active-1)):
+            m.d.dvi += self.x.eq(self.h_reset)
+            with m.If(self.y == (timings.v_active-1)):
+                m.d.dvi += self.y.eq(self.v_reset)
+            with m.Else():
+                m.d.dvi += self.y.eq(self.y+1)
+        with m.Else():
+            m.d.dvi += self.x.eq(self.x+1)
 
-        # Encode color data with xor/xnor
-        enc_qm = Signal(9)
-
-        # First bit is unmodified
-        m.d.comb += enc_qm[0].eq(self.data_in[0])
-
-        # Generate the remaining bits using xor/xnor
-        for i in range(7):
-            m.d.comb += enc_qm[i+1].eq(Mux(
-                use_xnor,
-                enc_qm[i] ^ ~self.data_in[i+1],  # XNOR
-                enc_qm[i] ^ self.data_in[i+1]    # XOR
-            ))
-
-        # Set indicator bit
-        m.d.comb += enc_qm[8].eq(~use_xnor)
-
-        # Calculate disparity for DC balancing
-        ones = Signal(signed(5))
-        zeros = Signal(signed(5))
-        balance = Signal(signed(5))
-
-        # Count ones in encoded data
-        m.d.comb += ones.eq(
-            enc_qm[0] + enc_qm[1] + enc_qm[2] + enc_qm[3] +
-            enc_qm[4] + enc_qm[5] + enc_qm[6] + enc_qm[7]
-        )
-
-        # Calculate zeros and balance
+        # Generate all control signals based on index.
+        # Sync inversion must be handled before the PHY (see below).
         m.d.comb += [
-            zeros.eq(8 - ones),
-            balance.eq(ones - zeros)
+            self.ctrl.hsync.eq((self.x >= self.hs_start) & (self.x < self.hs_end)),
+            self.ctrl.vsync.eq((self.y > self.vs_start) &
+                               (self.y < self.vs_end) |
+                               ((self.y == self.vs_start) & (self.x >= self.hs_start)) |
+                               ((self.y == self.vs_end)   & (self.x < self.hs_start))),
+            self.ctrl.de.eq((self.x >= 0) & (self.y >= 0)),
+
         ]
 
-        # Main TMDS encoding process
-        with m.If(~self.de):
-            # Control data during blanking interval
-            with m.Switch(self.ctrl_in):
-                with m.Case(0b00):
-                    m.d.dvi += tmds_r.eq(0b1101010100)
-                with m.Case(0b01):
-                    m.d.dvi += tmds_r.eq(0b0010101011)
-                with m.Case(0b10):
-                    m.d.dvi += tmds_r.eq(0b0101010100)
-                with m.Default():
-                    m.d.dvi += tmds_r.eq(0b1010101011)
-            # Reset bias
-            m.d.dvi += bias.eq(0)
-
-        with m.Else():
-            # Pixel color data logic
-            with m.If((bias == 0) | (balance == 0)):
-                # No prior bias or disparity
-                with m.If(enc_qm[8] == 0):
-                    m.d.dvi += [
-                        tmds_r.eq(Cat(~enc_qm[0:8], Const(0b10, 2))),
-                        bias.eq(bias - balance)
-                    ]
-                with m.Else():
-                    m.d.dvi += [
-                        tmds_r.eq(Cat(enc_qm[0:8], Const(0b01, 2))),
-                        bias.eq(bias + balance)
-                    ]
-            with m.Elif(((bias > 0) & (balance > 0)) | ((bias < 0) & (balance < 0))):
-                m.d.dvi += [
-                    tmds_r.eq(Cat(~enc_qm[0:8], enc_qm[8], Const(1, 1))),
-                    bias.eq(bias + Cat(Const(0, 3), enc_qm[8], Const(0, 1)).as_signed() - balance)
-                ]
-            with m.Else():
-                m.d.dvi += [
-                    tmds_r.eq(Cat(enc_qm[0:8], enc_qm[8], Const(0, 1))),
-                    bias.eq(bias - Cat(Const(0, 3), ~enc_qm[8], Const(0, 1)).as_signed() + balance)
-                ]
+        # Apply inversion where required.
+        m.d.comb += [
+            self.ctrl_phy.de.eq(self.ctrl.de),
+            self.ctrl_phy.hsync.eq(self.ctrl.hsync ^ self.timings.h_sync_invert),
+            self.ctrl_phy.vsync.eq(self.ctrl.vsync ^ self.timings.h_sync_invert),
+        ]
 
         return m
 
 class DVIPHY(wiring.Component):
     """
-    DVI PHY, outputs currently ECP5-specific
-
-    Note: Assumes clock domains 'dvi' (pixel clock) and 'dvi5x' (5x pixel clock)
+    DVI PHY (serializer), DDR outputs currently ECP5-specific.
+    Assumes clock domains 'dvi' (pixel clock) and 'dvi5x' (5x pixel clock).
     """
-
     # Channel data and control inputs
-    de: In(1)          # Data enable (high when drawing)
+    de: In(1)           # Data enable (high when drawing)
     data_in_ch0: In(8)  # Channel 0 - data
     data_in_ch1: In(8)  # Channel 1 - data
     data_in_ch2: In(8)  # Channel 2 - data
@@ -146,9 +126,9 @@ class DVIPHY(wiring.Component):
         m = Module()
 
         # TMDS encoders for each channel
-        m.submodules.encode_ch0 = encode_ch0 = TMDSEncoder()
-        m.submodules.encode_ch1 = encode_ch1 = TMDSEncoder()
-        m.submodules.encode_ch2 = encode_ch2 = TMDSEncoder()
+        m.submodules.encode_ch0 = encode_ch0 = tmds.TMDSEncoder()
+        m.submodules.encode_ch1 = encode_ch1 = tmds.TMDSEncoder()
+        m.submodules.encode_ch2 = encode_ch2 = tmds.TMDSEncoder()
 
         tmds_ch0 = Signal(10)
         tmds_ch1 = Signal(10)
